# adv-net-AdvancedBackendDotNet/openapi.yaml
openapi: 3.0.3
info:
  title: Swagger Enrollee - OpenAPI 3.0
  description: |-
    Система "Абитуриент" для продвинутого курса по .NET
    Сваггер не дает забиндить несколько описаний одного кода ошибки :/
  version: 0.0.0.0.0.0.0.0.0.0.0.0.1
...
paths:
...
  /profile:
    get:
      tags:
        - profile
      summary: Получить информацию о профиле
      description: Получить информацию о профиле
      operationId: getProfile
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/profileResponseModel'          
        '400':
          description: Invalid data
        '401':
          description: User not logged in
        '404':
          description: User not found
      security:
        - enrollee_auth:
            - allScope
    put:
      tags:
        - profile
      summary: Редактировать информацию о профиле
      description: Редактировать информацию о профиле
      operationId: editProfile
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/profileResponseModel'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/messageResponseModel'          
        '400':
          description: Invalid data
        '401':
          description: User not logged in
      security:
        - enrollee_auth:
            - allScope
...
components:
  securitySchemes:
    enrollee_auth:
      type: oauth2
      flows:
        implicit:
          authorizationUrl: https://petstore3.swagger.io/oauth/authorize
          scopes:
            allScope: xdd
    api_key:
      type: apiKey
      name: api_key
      in: header
  schemas:
...
    messageResponseModel:
      type: object
      properties:
        message:
          type: string
    profileResponseModel:
      required:
        - name
        - surname
        - gender
        - birthDate
        - citizenship
        - phoneNumber
      type: object
      properties:
        name:
          type: string
        surname:
          type: string
        secondName:
          type: string
        gender:
          type: string
          enum:
            - male
            - female
        birthDate:
          type: string
          format: date
        citizenship:
          type: string
        phoneNumber:
          type: string
...